% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{readLog}
\alias{readLog}
\title{Read log file generated by \code{ForceLogger}}
\usage{
readLog(file)
}
\arguments{
\item{file}{A path of log file}
}
\description{
Read log file generated by \code{ForceLogger} and return structured log as \code{data.frame}.
}

