% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{SETUP_FORCE_LOGGER}
\alias{SETUP_FORCE_LOGGER}
\title{Set the threshold for FORCE logger & Set the file path to write log}
\usage{
SETUP_FORCE_LOGGER(threshold = "INFO", logFilePath = "test.log")
}
\arguments{
\item{threshold}{A character.
One of \code{"TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL"}.
Default is \code{"FATAL"}.}

\item{logFilePath}{A file path. Default is \code{NULL}. If it is \code{NULL}, a log file is not written.}
}
\description{
The \code{threshold} affects the visibility of FORCE loggers.
}

